app/requirements.txt
uvicorn==0.34.0
fastapi==0.115.8
azure-identity==1.20.0
azure-cosmos==4.9.0